Storage terminology and requirements
Addresses issues #51 and #58

The text about updating is not lost, it's just part of PR #80
so it's needed in the terminology section.

It turned out to be too hard to reference SUIT manifest for definitions
of the trust anchor and trust anchor store because the TEEP arch already
had a definition in, and because TEEPs use of trust anchors itself is
independent of its use of a SUIT manifest.  Hence copied in text.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>