Remove normative language
Addresses issue #69

Removal of some uses of normative language in attestation text was
covered by PR #81.  This PR covers all uses not already part of
the other PR.

Any discussion of OCSP is left to the protocol doc, since OTrPv1
directly depended on OCSP and the new TEEP protocol is still in flux.
Hence removed mention of OCSP in particualr from the arch doc.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>