Address comments from Tiru and part 1 of Daniel Migault's
Addresses Tiru's comments on issues #178, #183, and #185.

Extra text in the definition of Device is removed because it confused
Daniel, and because it's now redundant with text in section 5.1
(Trust Anchors in a TEEP Agent) now that that section has been
revised.

Several other editorial nits pointed out by Daniel are addressed.
Other changes for Daniel's comments will be in a separate PR.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>