Revise compromised and hostile TAM threat sections
Updated compromised TAM threat section: a TAM cannot install arbitrary TAs. There is the second Trust Anchors in the TEE that validates TA signer to be trust worthy or not. Also mentioned the risk in TAM side by leaking data to an adverse actor.

In the case of a hostile TAM, the mitigation focus changes to stop using the TAM, and switch to use another one.